pyproject.toml
[project]
name = "jobhunter"
version = "1.0.0"
description = "Track new job postings from target companies (DoorDash, Apple, and more)."
requires-python = ">=3.10"
authors = [{ name = "Raghav Kachroo" }]
dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.0",
]

[project.scripts]
jobhunter = "jobhunter.app:main"

[tool.setuptools.packages.find]
include = ["jobhunter*"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
